Remove KZG-precompiles specific logic from EVM core/utils
